fix(v0.2.1): Fix compilation errors from design rework
Resolve 24 build errors found during flutter analyze/build after the
v0.2.1 design rework (commit 3992502). All fixes verified against
actual SDK model files:

- Add SDK import + hide conflicting DecisionList in decisions_screen
- Fix Duration/String and String/Color type mismatches in decision_sidebar_panel
- Regenerate stale .g.dart files for experiments and teams providers
- Hide SDK hypothesisListProvider to resolve ambiguous import
- Fix positional→named arg in showCreateHypothesisDialog call
- Fix TeamMemberRole enum constants (lead→LEAD, owner→OWNER, member→MEMBER)
- Fix String→Color conversion for stakeholder performanceColor
- Fix num→int cast for pendingDecisionCount accumulation
- Fix List<Hypothesis> access (removed .content on non-paginated response)

+++ lib/features/teams/screens/teams_screen.dart
@@ -38,7 +38,7 @@ class TeamsScreen extends ConsumerWidget {
                   // Calculate pending decisions across team
                   int totalPendingDecisions = 0;
                   for (final stakeholder in stakeholders) {
-                    totalPendingDecisions += stakeholder.pendingDecisionCount;
+                    totalPendingDecisions += stakeholder.pendingDecisionCount as int;
                   }
 
                   return Text(
